+++ climada/entity/exposures/test/test_base.py
@@ -321,6 +321,7 @@ class TestIO(unittest.TestCase):
         self.assertDictEqual(exp_df.meta, exp_read.meta)
         self.assertTrue(u_coord.equal_crs(exp_df.crs, exp_read.crs))
         self.assertTrue(u_coord.equal_crs(exp_df.gdf.crs, exp_read.gdf.crs))
+        self.assertEqual(exp_df.description, exp_read.description)
         np.testing.assert_array_equal(exp_df.gdf.latitude.values, exp_read.gdf.latitude.values)
         np.testing.assert_array_equal(exp_df.gdf.longitude.values, exp_read.gdf.longitude.values)
         np.testing.assert_array_equal(exp_df.gdf.value.values, exp_read.gdf.value.values)
@@ -428,6 +429,7 @@ class TestGeoDFFuncs(unittest.TestCase):
         self.assertTrue(u_coord.equal_crs(exp_copy.crs, exp.crs))
         self.assertEqual(exp_copy.ref_year, exp.ref_year)
         self.assertEqual(exp_copy.value_unit, exp.value_unit)
+        self.assertEqual(exp_copy.description, exp.description)
         np.testing.assert_array_equal(exp_copy.gdf.latitude.values, exp.gdf.latitude.values)
         np.testing.assert_array_equal(exp_copy.gdf.longitude.values, exp.gdf.longitude.values)
 
@@ -441,6 +443,7 @@ class TestGeoDFFuncs(unittest.TestCase):
         self.assertTrue(u_coord.equal_crs(exp.crs, 'epsg:3395'))
         self.assertEqual(exp.ref_year, DEF_REF_YEAR)
         self.assertEqual(exp.value_unit, DEF_VALUE_UNIT)
+        self.assertEqual(exp.description, None)
 
     def test_to_crs_pass(self):
         """Test to_crs function copy."""
@@ -453,6 +456,7 @@ class TestGeoDFFuncs(unittest.TestCase):
         self.assertTrue(u_coord.equal_crs(exp_tr.crs, 'epsg:3395'))
         self.assertEqual(exp_tr.ref_year, DEF_REF_YEAR)
         self.assertEqual(exp_tr.value_unit, DEF_VALUE_UNIT)
+        self.assertEqual(exp_tr.description, None)
 
     def test_constructor_pass(self):
         """Test initialization with input GeoDataFrame"""